feat: auto-rebuilding modules & FS watching
- Rename start command to autoreload.

- Base autoreload for a given module on its services' implied deploy
tasks (or just build, if the module doesn't define any services).

- Add version property and associated calculation methods to build &
deploy tasks, and modified getKey to incorporate versions when
appropriate.

- In connection with this change, build & deploy tasks are
now created via an async "factory" method, to ensure that the version
property for those task types has been computed before their getKey
method can be called.

- Modified TaskGraph to support inheritance/redo: If a task with the
same baseKey already already exists in the graph and is in progress
when a new task is added, the new task is set to inherit from the
in-progress task (see task-graph.ts for details).