refactor: major overhaul to plugin architecture
Plugins are now registered and loaded separately, and are now defined
as modules with an exported `gardenPlugin` function.

BREAKING CHANGE:
This includes some changes to the project schema and how it is resolved,
as well as how the main `Garden` class is instantiated. The `Garden`
class is now called with an environment name, which is then fixed for
the session. The env configuration is resolved by merging the specific
environment configuration with a global configuration specified on the
new `global` key in the project config. The schema for the `providers`
key also different - its keys should now match plugin names, and
contain configuration for those plugins.