feat: hot-reload functionality for local k8s
When the local-kubernetes provider is used, container modules can now be
configured to hot-reload their running services when the module's
sources change (i.e. without redeploying).

Services that can be run with an FS watcher that automatically applies
changed sources to the running application (e.g. nodemon and many other
web app frameworks) are a natural fit for this feature.