refactor(go-cli): use a dedicated sync container (wip)
With this change, garden mounts the project dir into a dedicated sync
container _if_ the command requires file watching. Within the sync
container, the project dir is copied into a named volumed and a
one directional unison process is started that keeps the named volume
synced with the project dir. Finally the garden-service container
runs with the named volume mounted.

If the command does **not** require file watching, the project dir is
bind mounted straight into the garden-service container as before.

This is all very much experimental.