feat: support ignorefiles in subdirectories
Previously, getIgnorer only considered .gitignore and .gardenignore
files located in a project's root directory.

We now scan for .gitignore and .gardenignore files (referred to as
ignorefiles) in all subdirectories of the project's root directory.

Files nested within a subdirectory (of project root) that contains
one or more ignorefiles will only be matched against the rules found in
those ignorefiles.

This is done recursively - a file is only matched against the most
deeply nested set of ignorefiles located in its parent directories
(including the one it's enclosed in, if relevant).

test: added tests for wildcard rules

feat: support ignorefiles in subdirectories