feat: add support for flat config style
Projects and modules can now be configured in a flat style.

That is, not nested under the project/module key, but with the entity type
indicated by the new `kind` key, analogously to the YAML syntax for k8s
object definitions.

Also changed the (currently unused) schema version string format to
`garden.io/[version]`.