fix: don't emit taskPending if task is skipped
This fixes an issue in TaskGraph where taskPending events were emitted
for tasks with up-to-date cached results. This led to the task graph UI
on the dashboard displaying the nodes corresponding to skipped tasks in
a permanent pending state.

This was addressed by only emitting the event if the task was, in fact,
added to TaskGraph's index.