fix(build): remove deleted files/dirs during sync
Before this fix, removing non-ignored files or folders from a module
during the execution of a watch-mode command would not result in those
files/folders being removed from the module's build directory.

This was fixed by adding the --delete option to the rsync call used by
BuildDir, and moving the .garden-version file from the module's build
directory to a new subdirectory of .garden,
build-metadata/[module-name].

In general, the build-metadata directory can be used by plugins to
persistently store build-related metadata.