feat(k8s): cache task results
Added caching for task results. Similarly to the caching for test runs
that was already in place, these are stored by version as ConfigMap-s.

The version used to index these results includes the task's module's
current version, along with the module versions of the task's
runtime dependencies (if any). This is the same formula that is used for
computing test task versions.

Currently, tasks always execute, regardless of whether they have cached
results for the current version, since we need to finish other changes
that will provide dependency results to the process methods of task
classes before the caching semantics can be finalized for tasks, in
relation to dependency calculations.

A versioning bug was also fixed for the test result caching
functionality. Here, the version used was just the test config's
module's version, whereas now the corresponding test task's version
is used to index it instead (the same as is now also used to cache
task results).