fix: include resolved config in module version
Module versions now also include a serialized version of the resolved
module config (hashed together with versions / dirty timestamps).

This fixes an issue where changes to project variables used in a
module's config would not result in the module's version changing when
recomputed.

This was the case because the version was based only on the latest
commit and dirty timestamp of the module in question, and those of its
dependencies, which doesn't factor in changes to project-level
variables.