docs: use flat config style in example projects
Also renamed a few example projects (and adjusted links to them in the
docs accordingly):

hello-world -> openfaas
simple-project -> demo-project
simple-project-start -> demo-project-start

The hello-world.md guide was removed, since it's mostly a duplicate of
the material in demo-project.md (previously simple-project.md).

In a later PR, we'll add some how-to sections that will cover what is
omitted here from the old hello-world guide, as well as adding new
material in that format (instead of wrapping it in a guide as was done
in hello-world.md).