feat(core): add module include field and use content hash for versions
This changes the version handling to no longer rely on git commit hashes
but instead use content hashes on all included files.

Also added is the `include` field on modules, which allows for control
over which files are considered part of the module sources.

Further, this improves handling of .gardenignore files, so they now
work in subdirectories as well as the project root.

Note: .gitignore files no longer factor in when evaluating module
versions. Use .gardenignore files instead (same format) to exclude files
from file watching and versioning in Garden.

Closes #328
Closes #603