fix(k8s): deduplicate ns creation during init
Before this fix, environment initialization for the OpenFAAS plugin
would fail because a second call to ensureNamespace was made before the
first call (with the same namespace name) was able to register that the
namespace had been created. This created a race condition.

This was fixed by maintaining a "pending" status to deduplicate such
requests.