feat(container): add command option
Added `command` and `hotReloadCommand` fields to container module
configs to go with `args` and `hotReloadArgs`  (on services, test
configs and tasks).

For the k8s plugins, `command` corresponds to the `command` field on k8s
podspecs, i.e. it's the "entrypoint" which `args` are passed to.

This is useful e.g. when some other entrypoint than `/bin/sh -c` is
desired.