improvement(k8s): more robust and useful deployment status checks
This combines a general refactor of our workload status checks, more
robust checking of related resource statuses (in order to fail fast on
common errors like ImagePullBackOff and CrashLoopBackOff), as well as an
improvement on the level of information given when errors do come up.

This should aid considerably in debugging deployment issues.

Closes #863