fix(core): chokidar watcher on mac could segfault after reloading configs
Turns out chokidar and the native fsevents module don't like it when multiple
instances are created in a single process, even if one is closed before the
other starts. Fixed it by making a global instance and reconfiguring on
reload instead. See https://github.com/fsevents/fsevents/issues/273 for
a discussion on this issue.