improvement(config): explicitly validate sub-paths when applicable
We add a `path` rule, with options to disallow absolute paths or parent
paths, i.e. stepping up a directory tree, in order to ensure that some
paths are sub-paths of a project or module.

This required a refactor where we use a custom Joi instance across our
codebase, hence the size of the change.

This is done both for general hardening, and in preparation of #853 and
other similar configuration options.