fix(task-graph): fix task deduplication
Fixed a regression that allowed more than one pending (i.e. not
in-progress) task to be added to `TaskGraph` for a given `key`.

This happened when the add & process queues were merged into a single
op queue (for valid reasons), and could cause repeated file changes
to result in unnecessarily repeated builds, deploys etc.

After this fix, at most one task can be pending for a given `key`.