fix(k8s): make sure Service Endpoints are ready at end of status checks
This step is put in to give the cluster a moment to update its network
routing. For example, when a Deployment passes its health check,
Kubernetes doesn't instantly route Service traffic to it (particularly if
the master is under high load). We need to account for this so that
dependant tasks, tests and services can reliably run after the resource
status check resolves.