chore(release): bump version to v0.10.1-0
Note, the version in garden-service/package.json and
garden-service/package-lock.json is v0.10.1 instead of v0.10.1-0. This
is due to a bug in lerna that causes lerna bootstrap to fail when the
version number is a pre-release. This shouldn't matter since we'll
update the version anyway when we make the proper release.