improvement: updated the get debug-info command
The old version of the command might include sensitive data in the output file.
This change introduces more explicit logging to show what's been extracted.
Changes the behaviour of the command when it comes to export data from the kubernetes cluster:
by default the command will only dump data from the garden-system and garden-system--metadata namespaces.
If the user uses the new flat --include-project, the command will dump the configuraition for the entire project as well