fix(watcher): native fsevents were not used in dist build on macOS
This caused a tremendous slowdown and high CPU usage on Mac. The fix is
essentially to use a newer version of Pkg for macOS (but not for other
builds because the newer version breaks other platforms, yay), as well
as to update the fsevents binary.

I've also put in an assertion to try to prevent this happening again
without us noticing.