refactor(plugins): make plugin definition interface more intuitive
This is to prepare for more changes related to the plugin SDK, and to
overall make the interface more intuitive. For example, we now
distinguish more explicitly between creating and extending module types.
Also added a bit of validation when loading plugins (see the added
tests).

There should be no visible change in usage or operation.