refactor(dashboard): pass deps to useEffect hook
Previously we passed an empty array as a dependency to the useEffect
hooks that were responsible for data fetching to ensure they only run
once on mount. This is considered a bad practice so now we've refactored the
data fetching logic so that we can explicitly pass all dependencies.

We do this by removing the action creators and calling the load
functions directly from the container components.