chore(dashboard): fix linting some more
We now use the @typescript-eslint/tslint plugin to import the dashboard
tslint config which extends the root level config. This allows us to use
all our tslint rules inside the dashboard .eslintrc.js file. We need
this because the dashboard does not support using tslint directly.

Tslint rules that are imported like this are not autofixable. However, a
lot of the rules we need aren't available with @typescript-eslint and
we'd rather get a non-fixable lint error as opposed to no error at all.