docs: ensure check-docs script checks internal header links
markdown-check-link doesn't check internal header links. This is a
known issue (https://github.com/tcort/markdown-link-check/issues/65).

We therefore add a new tool to check for internal links and limit
markdown-check-link to external links.

The new tool is from the remark family of markdown tools and expects
anchor links to not include any special characters so we also had to
make appropriate changes to the docs generator functions.

Finally, this commit fixes broken links discovered because of these
fixes.