refactor(plugin): implement module type inheritance and polymorphism
This expands on the work in #1256. We now allow inheritance on module
type declarations, and for plugins that support a base module type to
transparently also support any module types that inherit from it.

Case in point, the `maven-container` module type now has `container`
as a base, and plugins that support `container` modules (e.g. the
`kubernetes` plugin) now also support `maven-container` modules without
having to be specifically aware of them.

We can use this to make an array of different "special-case" module
types that then retain compatibility with plugins that work with their
base.