refactor(core): allow plugins to augment the module graph
Providers are now able to augment the stack graph using a
new `augmentGraph` handler, which is called in dependency
order for each provider that specifies it. This allows
providers to add modules (and by extension services, tasks,
tests), and to add dependency relations between existing
modules and services/tasks.

See the new augmentGraph plugin handler for details on usage.

Also made a small refactor to no longer require a build
handler for every module type, and to avoid unnecessary
build tasks when a module doesn't require building.