feat: hadolint provider
The `hadolint` provider is a fairly simple plugin that uses
the new `augmentGraph` handler. Simply add the `hadolint` 
provider to your project config, and a linting test will
be created for each `container` module in the project. See
the `hadolint` provider and `hadolint` module type reference
docs for details and more usage options.