improvement(container): automatically set include field based on config
This avoids common issues where multiple module sources overlap, or
Garden build contexts accidentally include a bunch of unnecessary files.

We parse the Dockerfile if the module has one and extract the paths
referenced in the ADD and COPY fields there. If there is no Dockerfile,
we automatically set include to an empty list.