fix(build): ensure that exec modules are rebuilt on when garden is run
Prior to this commit, garden would skip building exec modules that had been built during a previous garden run, if none of the source files in the module changed.

While this is a good optimisation, it caused problems because staging would still happen, wiping out any previously built artifacts which other modules might have depended on.  Also, there is no way to guarantee/specify at the module level that there are no outside dependencies that aren't captured by the module's sources.

It is planned to properly support this in the future, but removing this optimisation for now.