refactor: update Joi library
This turned out to be quite involved, but it's good to just get it over
with before moving on to any further improvements to our validation.

I had to fork and use the latest master (the fork being to make sure
we have a consistent version) so that we get a recent fix in there. We
should get rid of that once the next release (after 16.1.8) is out.