feat(core): remember IP per service for forwards on supported platforms
Unfortunately macOS does not support binding to local addresses other
than 127.0.0.1 without explicit aliasing which requires sudo privileges.
Other platforms and Linux setups may have this limitation as well, so
we gracefully fall back when we're not allowed to bind to our preferred
addresses.

This is a fairly low-key feature in and of itself, but includes the
major addition of a user-wide SQlite database, which we can use for a
number of other things. The PR includes the basic setup to build
features that leverage persistent local state, and through the ORM we
can later consider supporting other types of databases.