fix(k8s): reconnect port-forwards automatically
kubectl port-forward errors are now caught and the process killed, and
the Garden proxy now handles such errors. This handles cases where Pods
are terminated for whatever reason, including after re-deployment.

It would be possible to further improve this by moving to use
the API directly instead of kubectl (which I spiked and looks like quite
a bit of work), and potentially to somehow signal to the proxy that the
underlying port forward has been terminated. It works fairly well as-is
though, so I'll leave those improvements for later.

Fixes #1345