fix(k8s): play nice with Helm 2 (Tiller) when users still need it
Users may need Tiller for non-Garden related reasons. We now no longer
automatically remove Tiller from project namespaces (but we do for
garden-system), and instead provide a command to do the cleanup and a
prompt for the user to run it.

NOTE: We need to update the 0.11 release notes to reflect this!