fix(helm): only hot reload serviceResource
Before this fix, the sync volume would be attached to all containers in
the target resource when using hot reloading with Helm modules, which is
not desirable e.g. when using user-specified sidecar containers.

Now, the sync volume is only attached to the container identified by
serviceResource.containerName or the Helm module's name.