fix(core): allow unknown in task & test results
This adds a bit of future-proofing for schema changes to test and task
results.

If unknown fields are not allowed in task and test results, new fields
stored on results generated by newer versions of the framework
may cause errors when read by older versions which do not recognize
the new fields.