chore: fix doc links
These links are from the auto-generated docs. So they get generated for
tag master but replaced by the release script, which breaks the
check-docs CI check. Fixing manually now so that we can merge this but
we need to a more permanent fix. E.g. by replacing the links in
garden-service/src as well or by setting the correct tag in the
generate-docs logic.