fix(k8s): tasks and tests would sometimes return empty logs
This was due to kubectl sometimes failing to attach to Pods. We now
explicitly fetch the logs, with a bit of trickery to be able to do so
when copying artifacts out of the container after running the test/task
command.

This should also address the flaky integ tests we've been seeing in
minikube.