feat(template): add optional suffix, to allow undefined values
In some cases, you may want to provide configuration values only for
certain cases, e.g. only for specific environments. By default, an
error is thrown when a template string resolves to an undefined value,
but you can explicitly allow that by adding a `?` after the template.

Example:

```yaml
kind: Project
...
providers:
  - name: kubernetes
    kubeconfig: ${var.kubeconfig}?
  ...
```

This is useful when you don't want to provide _any_ value unless one is
explicitly set, effectively falling back to whichever the default is
for the field in question.