fix(core): fix concurrency bug in task graph
This fixes a race condition discovered in the recently introduced
batch-oriented concurrency changes to the task graph.

Here, we store each task node's original dependencies directly on the
node itself, and do away with the `taskDependencyCache` that was
previously used (and seemed to be the cause of the race condition).

We used the opportunity to do some refactoring and cleaning up in the
`TaskGraph` class while we were at it.