feat: add persistentvolumeclaim module type and volumes for containers
This makes it possible to mount persistent volumes in container modules,
including in tasks and tests. With the right storage classes in place,
you can even share volumes between container services, tasks and tests.
These volumes can of course also be referenced in `kubernetes` and
`helm` modules.

See the added section to the container modules guide for details on
how to use the new features.

As a small bonus, in order to be able to use the upstream
PersistentVolumeClaim schema, I've added a
`joi.customObject().jsonSchema()` type, which allows us weave JSON
schemas into our existing Joi schemas. My hope is that we can gradually
get rid of most usages of Joi schemas, at least in our plugin APIs.