fix(task-graph): don't re-run failed tasks
After this change, a task will not be re-run if it has force = false and
another task with the same key and version has already thrown an error
while running (and thus had its result cached in the graph's
`ResultCache`).

This means that we treat task results as cached regardless of whether
the task succeeded or not: The semantic is that a task with force =
false should always succeed with the same result (or fail) for a given
key + version.

When this is not the case, force = true should be used.

This prevents e.g. time-consuming tasks from repeatedly running and
failing in the same fashion when repeatedly added to the task graph
following watch changes.