fix(core): fix to task deps for DeployTask
`DeployTask`'s `getDependencies` method now always returns task
dependencies initialized with `force = false`.

The underlying semantic is that unless the task dependency's module
sources have changed, there's no need to re-run them when
force-deploying the dependant service.

I.e. the decision to run or not run the task dependency is delegated to
the `TaskTask` in the normal way.