UPDATED quick-start.md
It is confusing explaining that the `--env=remote` is only needed when using a remote OR edit the garden.yaml. THEN the first command has the `env` flag stated but none of the following commands do. This confused me for hours. Removing the flag and going start-to-finish now makes a bit more sense.