fix(core): fix to dependency logic in dev command
Before this fix, if a service was deployed with hot reloading via the
dev command, and a test spec had a runtime dependency on that service,
the service would initially be deployed without hot reloading enabled.

This would result in subsequent hot reloads for that service failing.

Also added a test case checking for this behavior.