fix(commands): don't use default port in serve cmd
Before this commit, the `garden serve` command would use
`DEFAULT_PORT` as the default value for its `port` option.

If another command was already running and serving on
`DEFAULT_PORT` when `garden serve` was called, this would result in an
error.

This was fixed by deferring to the standard random port selection logic
we use for other commands by skipping the default value for this command
option.