chore: add profiler for performance analysis
Adding some tooling to analyze how long specific parts of our code take
to complete. A summary is printed at the end of unit and integ test
runs, which should help us where we can make optimizations.

Note that the syntax for wrapping/profiling individual functions is
slightly odd, simply because there isn't any native support for
function decorators in JS. Other ideas are welcome, but I think this
is fine for now.