feat(config): allow any objects and arrays in project variables
This lifts a limitation that is no longer necessary, and adds quite a
bit of flexibility to our variables mechanism. You can now put any
object or array in your project variables (both global and
environment-specific), and template strings can also output any object
or array.

See the updated docs on variables and templating for more details, as
well as the tests.