fix(core): fix watch task logic for sourceModules
Before this fix, if a service with a sourceModule was deployed with hot
reloading enabled, a build task for its source module would be added
when its sources change (in addition to the hot reload task).

This was not the desired behavior, and would (e.g. for helm modules
using a container module as a sourceModule) result in unnecessary
rebuilds of the source module.