fix(terraform): fix stdin handling in commands
Before this fix, stdin was not being handled correctly for the terraform
plugin's commands.

This was fixed by not setting setRawMode to true on the child process'
stdin.

We added the rawMode option to SpawnParams to facilitate other use cases
that need interactive input, but where rawMode doesn't play nicely with
stdin for the underlying child process.