feat(core): support namespaces natively at the framework level
This makes namespaces an explicit framework feature, which can be
reasoned about and leveraged by providers and external tools.

The change is backwards compatible, and doesn't require any usage or
configuration change.

Documentation and example updates will follow in a separate PR, to avoid
confusion with the current release version.