feat(core): workflows for sequencing commands
Workflows are a new framework primitive. They specify a list of
`steps`, each of which specifies a Garden command to be run.

They also specify `triggers`, which facilitate automatic triggering of
workflows based on VCS events.

`Workflow` configs are a new top-level config type, along with
`Project` and `Module`.

The `run workflow` command can be used to run a workflow by name.