perf(core): cache provider statuses for faster successive startup
We now cache provider statuses for a default of one hour. You can run
any command with `--force-refresh` to skip the caching, and can override
the cache duration with the `GARDEN_CACHE_TTL=<seconds>` environment
variable.

This substantially improves command execution times when run in
succession, which will be very useful for day-to-day usage, as well CI
performance.

Closes #1824