fix(helm): allow runtime templates in helm module values field
We now render the values we feed into Helm on every render, instead of
only at build-time. This unblocks the usage of `${runtime.*}` template
strings, which in turns unblocks several common use-cases, including
feeding in outputs from Terraform modules.

This required making a specific conftest module type that is Helm aware.
Once we've implemented a render/plan phase in our framework, we should
be able to get rid of that specific implementation.

Fixes #1881