fix(kubernetes-module): fix namespace handling
Before this fix, we were using the module/service name as the annotation
value for all manifests/resources.

This didn't work for `Namespace` resources in particular, since
deploying a `kubernetes` module that included a `Namespace` manifest
would result in namespaces with different names (but deployed from the
same module, e.g. by another user in the same cluster) being deleted.

The same applied to the `delete service` command when used in the same
context.

We now annotate `Namespace` manifests/resources with the namespace name,
which fixes this problem.