docs: fix links on intro page
The links on the intro paged used to point to the README.md file of each
section. However, those READMEs are not included in the summary file
(docs/README.md) and therefore won't work on Gitbook. If we want to link
to the section READMEs, we need to explicitly add them to each section.