chore: ensure git user config is set in GitHub workflow
The git commit step in our GitHub action started failing with exit code
128 because the user name and email weren't set in the git config. Not
sure why this stopped working all of a sudden, perhaps changes to the
underlying git version or in how values are propagated between steps in
GitHub actions. In any case, setting these values with the --global flag
fixed the issue.