feat(dashboard): new garden dashboard command
The new `garden dashboard` command is designed to run standalone, and
receive events from other Garden processes. Any other process running
with the same project, environment and namespace will pipe events to
this process when it's running.

All active Garden processes are now cataloged in a table in our sqlite
database, and this table is polled for active processes during all
CLI execution. This will no doubt be handy for other stuff later.

The dashboard server adds an authenticated `/events` endpoint, which
uses the same API as the same endpoint in GE, and much of the same
event streaming code is used.

This should make the dashboard much more usable in day-to-day dev work,
and I aim to continue work in that direction.

There are some further improvements needed to make the dashboard better.
One of which is to fix reconnection issues, and we also need to make it
more generally useful by leaning on its pluggability, making more handy
providers which embed pages in the dashboard that will be useful for
development.

Notably, this improvement is also an important step in embedding
dashboards in the GE UI. This should be fairly easy now, since the
dashboard can receive events in the exact same shape as the GE event
store holds.

Once we feel this is stable, and we have a few more steps covered, I
will make a separate PR with a detailed usage guide.