feat(dashboard): add octant provider and dashboard integration
This adds an `octant` provider, that you can drop into any Garden
project that includes a `kubernetes`/`local-kubernetes` provider:

```
...
providers
  - name: octant
...
```

This will add an Octant tab to your dashboard, which will be loaded
on-demand, pointed at the context and namespace of your project.
You also get a `garden tools octant` command if you'd like to run it
standalone.

Most of the work in the PR is to do with the mechanism for providers to
add dashboard pages, which is now easy to do and can be done in a
dynamic fashion.

I also added a toggle for provider tools, to indicate whether they
should be pre-fetched as part of our container builds.