fix(core): undefined tool error in container module derivatives
This needed a fairly large refactor, but it's good to get it out of the
way. We now fetch tools from the PluginContext, instead of picking them
off providers, which was more brittle and harder to reason about.

We do still need to do a bit more work on the `maven-container` module
type, but that's for another PR.

Fixes #2002