improvement(cli): catch OOM errors and exit with helpful error message
This is one of a series of improvements I'm working on, to warn users
about very large projects that probably need better configuration of
includes/excludes/ignores.

This catches the "nuclear" case of running out of heap space, exiting
with a standard code of 137, and printing an error with a link to docs.