feat(container): add deployment-image-id module output key
As well as a corresponding local-image-id key.

This will seem like a large commit for a small feature, but this
involved adding a new plugin handler to allow referencing the module
version in module outputs, which was previously not possible.

We add a new getModuleOutputs plugin handler, and no longer return
outputs in the configureModule handler.