chore: fix test-release script
This script stopped working after the yarn update, due to the changed
version handling in our `package.json` files.

The prerelease suffix is not included in the version used for core, and
also not in the file name of the release archive on GitHub. We've now
added some logic to handle that.