fix(workflows): forbid use of global options
Global options (such as `--env` or `-l`) have no effect when used in
workflow step commands, since the step commands share a Garden instance
with the enclosing `run workflow` command.

To prevent user confusion, we now validate that global options are not
used in workflow step command specs, and throw an error if any are
detected.