fix(cli): less noisy error rendering for workflows
Previously, we were logging the details of failed tasks (along with
their successfully executed dependencies, if any) when a workflow
failed.

We've now simplified this to only render the error messages by default,
and rendering the details of the failed tasks only (without
dependencies) for log levels `debug` and higher.

The output logged to `error.log` has been streamlined in the same way,
omitting details of successful dependencies of failed tasks.