improvement(core): experimental build stage implementation without rsync
This implements the one-way build staging process entirely without
rsync. It's not trivial to get right, turns out, so for now this has
to be explicitly enabled by setting GARDEN_EXPERIMENT_BUILD_STAGE=true.

I've configured the e2e tests in CI to use the new mechanism, which
should help to surface issues, along with IRL usage testing.

We'll keep it that way until the next breaking release. Users who have
issues with the rsync-based implementation should give this a try and
by all means file issues and provide feedback.