improvement(core): more efficient file scanning with multiple ignores
This fixes a performance and memory consumption issue when specifying
multiple `dotIgnoreFiles`.

We do this using a streaming set intersection, which avoids storing
unnecessary data in memory during the scan, relying on a sorted file
order coming from git.

It's not exactly a trivial implementation but this is only applied when
multiple .ignore files are configured, which is anyway a problematic
routine prone to thrashing the heap.