improvement(dashboard): better Stack Graph layout
This replaces the previous dagre-based layout with Reaflow, which uses
ELK for the algorithm. This makes the layout stable for every run,
and uses a layered horizontal layout which scales much better with a
large number of nodes in the graph.

Reaflow also makes the code better fitted to React, so this should be
easier to work with as we do further iterations and improvements.