chore: use TS project references for (much) more efficient compilation
Also switched to use incremental compilation. This makes a pretty big
difference in my quick tests when developing.

One thing to look out for is that *we should now always run `yarn build`
and `yarn dev` from the project root* and not from the `core` directory,
which we used to do because it could be much more efficient.