feat: dev mode with much better container syncing
Introduces _dev mode_, which is a new, mutagen-based hot reloading
implementation.

Dev mode is currently flagged as experimental, but eventually the plan
is to deprecate the old rsync-based hot reloading in favor of dev mode.

Dev mode doesn't yet support a `postSyncCommand`. We should follow up
with that in a separate PR.