feat(k8s): allow setting annotations and labels on project namespace
You can now configure both annotations and labels on namespaces. The
`namespace` field on the `kubernetes` provider now also accepts an
object instead of just a string, so you can specify the desired
annotations and/or labels there.

These will be applied when creating namespaces, and Garden will also
attempt to ensure those annotations and labels are on existing
namespaces (emitting a warning if e.g. not allowed to modify the
namespace).