feat(config): add ${command.name} and ${command.params} for templating
This allows users to template by the current command context in some
interesting ways. A simple example would be template based on whether
the test command is being run.

A more complex example would be something like
`${command.params contains 'hot' && command.params.hot contains 'my-service'}`,
basically checking for a specific parameter.

Later I'd like to make this sort of thing more elegant in terms of
syntax, but it does already work nicely if used with a bit of care.