fix(core): potential edge-case issue with versions and generateFiles
This is a little bit of a shot in the dark, looking to solve an issue
reported on Slack
(https://kubernetes.slack.com/archives/CKM7CP8P9/p1614599864049600), but
it's in any case a sane thing to do. Basically we now ensure that a new
tree scan is performed after manipulating files in a particular module
path.